// services/web/src/api/schema.d.ts
export interface components {
    schemas: {
        /** ChatCompletionRequest */
        ChatCompletionRequest: {
            /** Model */
            model?: string | null;
            /** Messages */
            messages: components["schemas"]["ChatMessage"][];
            /** Max Tokens */
            max_tokens?: number | null;
            /** Temperature */
            temperature?: number | null;
            /** Top P */
            top_p?: number | null;
            /**
             * Stream
             * @default false
             */
            stream: boolean;
            /** Stop */
            stop?: string[] | null;
        };
        /** ChatCompletionResponse */
        ChatCompletionResponse: {
            /** Id */
            id: string;
            /**
             * Object
             * @default chat.completion
             * @constant
             */
            object: "chat.completion";
            /** Created */
            created: number;
            /** Model */
            model: string;
            /** Choices */
            choices: components["schemas"]["Choice"][];
            usage: components["schemas"]["Usage"];
        };
        /** ChatMessage */
        ChatMessage: {
            /**
             * Role
             * @enum {string}
             */
            role: "system" | "user" | "assistant";
            /** Content */
            content: string;
        };
        /** Choice */
        Choice: {
            /**
             * Index
             * @default 0
             */
            index: number;
            message: components["schemas"]["ChoiceMessage"];
            /**
             * Finish Reason
             * @default stop
             */
            finish_reason: string;
        };
        /** ChoiceMessage */
        ChoiceMessage: {
            /**
             * Role
             * @default assistant
             * @constant
             */
            role: "assistant";
            /** Content */
            content: string;
        };
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /** Usage */
        Usage: {
            /**
             * Prompt Tokens
             * @default 0
             */
            prompt_tokens: number;
            /**
             * Completion Tokens
             * @default 0
             */
            completion_tokens: number;
            /**
             * Total Tokens
             * @default 0
             */
            total_tokens: number;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
            /** Input */
            input?: unknown;
            /** Context */
            ctx?: Record<string, never>;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}